Improves container error message for easier debugging
In case the engine is not shut down correctly, config and data
container folders are left in their directories, making it impossible to
execute the same code again after a problematic exit. The new error
message containes the path to the folder where the container
directories reside.

In order to prevent people deleting directories that are still in use,
 a short informative message is also provided.

This addresses but does not solve issue IDSCETHZurich/rce#39